Refactor application into a component-based structure. This commit refactors the entire application from a single App.js file into a more organized and maintainable structure. Components, pages, and constants have been separated into their own directories and files. The main App.jsx now acts as a router, handling page navigation.
