+++ npm_output.log
@@ -1,0 +1,9 @@
+
+> tech-ai-learning-app@0.0.0 dev
+> vite
+
+
+  VITE v4.5.14  ready in 624 ms
+
+  ➜  Local:   http://localhost:5173/
+  ➜  Network: use --host to expose
