+++ src/components/MessageBox.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const MessageBox = () => (
+    <div id="message-box" className="message-box fixed top-4 right-4 z-50 p-4 rounded-lg shadow-lg transition-all duration-300 transform translate-x-full opacity-0">
+    </div>
+);
+
+export default MessageBox;
