+++ src/constants/translations.js
@@ -1,0 +1,84 @@
+export const translations = {
+    en: {
+        appName: "Tech & AI Learning",
+        signIn: "Sign In",
+        signUp: "Sign Up",
+        welcomeBack: "Welcome Back",
+        signInToAccount: "Sign in to your account",
+        createAccount: "Create an Account",
+        getStarted: "Sign up to get started",
+        emailPlaceholder: "Enter your email",
+        passwordPlaceholder: "Enter your password",
+        namePlaceholder: "Enter your full name",
+        magicLink: "Or use magic link",
+        alreadyAccount: "Already have an account?",
+        noAccount: "Don't have an account?",
+        chatHistory: "Chat History",
+        newChat: "New Chat",
+        noChatHistory: "No chat history",
+        welcomeChat: (name) => `Welcome ${name}! I'm ready to help you learn tech and AI.`,
+        chatPrompt: "Choose a learning topic or ask me anything.",
+        inputPlaceholder: "Ask about tech or AI...",
+        voicePrompt: "Tap microphone to speak",
+        languageToggle: "Canza zuwa Hausa",
+        languageHausa: "Hausa",
+        languageEnglish: "English",
+        loginSuccess: "Login successful! Redirecting to chat...",
+        signupSuccess: "Sign up successful! Redirecting to chat...",
+        fillAllFields: "Please fill out all fields.",
+        enterEmailPassword: "Please enter both email and password.",
+        sorryResponse: "Sorry, I could not generate a response.",
+        apiError: "I am currently unable to respond. Please try again later.",
+        voiceActivated: "Voice dictation activated. Start speaking...",
+        suggestTopics: "✨ Suggest Topics",
+        suggestionPrompt: (topics) => `Based on our conversation, you might also be interested in these topics:\n${topics.map(t => `* ${t}`).join('\n')}`,
+        noPreviousMessage: "Please send a message first to get suggestions.",
+        generatingSuggestions: "Generating suggestions...",
+        summarizeChat: "✨ Summarize Chat",
+        noChatToSummarize: "There are no messages to summarize.",
+        generatingSummary: "Generating summary...",
+        readAloud: "Read Aloud",
+        readingAloud: "Reading aloud...",
+    },
+    ha: {
+        appName: "Koyo da AI",
+        signIn: "Shiga",
+        signUp: "Bude Sabon",
+        welcomeBack: "Barka da dawowa",
+        signInToAccount: "Shiga asusunka",
+        createAccount: "Bude Sabon Asusunka",
+        getStarted: "Bude Sabon don farawa",
+        emailPlaceholder: "Shigar da email dinka",
+        passwordPlaceholder: "Shigar da kalmar sirri",
+        namePlaceholder: "Shigar da cikakken sunanka",
+        magicLink: "Ko amfani da hanyar sihiri",
+        alreadyAccount: "Kun riga kuna da asusu?",
+        noAccount: "Babu asusu?",
+        chatHistory: "Tarihin Tattaunawa",
+        newChat: "Sabon Tattaunawa",
+        noChatHistory: "Babu tarihin tattaunawa",
+        welcomeChat: (name) => `Barka da zuwa ${name}! Ina shirye don taimaka maka koyo game da fasaha da AI.`,
+        chatPrompt: "Zabi batun koyo ko tambaye ni komai.",
+        inputPlaceholder: "Tambaye game da fasaha ko AI...",
+        voicePrompt: "Danna microphone don magana",
+        languageToggle: "Switch to English",
+        languageHausa: "Hausa",
+        languageEnglish: "Turanci",
+        loginSuccess: "An yi nasarar shiga! Ana tura ku zuwa tattaunawa...",
+        signupSuccess: "An yi nasarar bude sabon! Ana tura ku zuwa tattaunawa...",
+        fillAllFields: "Don Allah cika duk filayen.",
+        enterEmailPassword: "Don Allah shigar da duka email da kalmar sirri.",
+        sorryResponse: "Sori, ban iya bayar da amsa ba.",
+        apiError: "A halin yanzu ba zan iya amsa ba. Don Allah gwada daga baya.",
+        voiceActivated: "An kunna umarnin murya. Fara magana...",
+        suggestTopics: "✨ Nemi batutuwa",
+        suggestionPrompt: (topics) => `Bisa ga tattaunawarmu, za ka iya kuma sha'awar waɗannan batutuwa:\n${topics.map(t => `* ${t}`).join('\n')}`,
+        noPreviousMessage: "Don Allah aika saƙo da farko don samun shawarwari.",
+        generatingSuggestions: "Ana samar da shawarwari...",
+        summarizeChat: "✨ Takaita Tattaunawa",
+        noChatToSummarize: "Babu saƙonni da za a takaita.",
+        generatingSummary: "Ana samar da takaitawa...",
+        readAloud: "Karanta da babbar murya",
+        readingAloud: "Ana karantawa...",
+    }
+};
